refactor: poll_message refactored to poll_messages (plus cargo fmt)
The main loop know, instead of matching a Some(message), it iterates
over the results of poll_messages, which collects the unhandled messages
and return an interator over them.
This way, no message is lost in the rare case of having two at the same
time; for example when a user presses a key at the same time a command
sends a message. (even tho commands are not yet implemented xd)